requirements.txt
# Web framework
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dotenv>=0.19.0
pytest>=6.2.5
httpx>=0.24.0
pyyaml>=6.0.1
python-multipart>=0.0.6
aiofiles>=0.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=1.4.0
alembic>=1.7.0
psycopg2-binary>=2.9.0
redis>=4.0.0
celery>=5.2.0
boto3>=1.26.0
openai>=1.0.0
tiktoken>=0.5.0
tenacity>=8.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
pdfminer.six>=20221105
pandas>=2.0.0
openpyxl>=3.1.2     # For Excel processing

# Logging
structlog>=23.1.0

# Environment variables
python-dotenv>=1.0.0

# LLM providers
anthropic>=0.5.0     # For Claude Sonnet 3.5

# LangChain
langchain>=0.3.21
langchain-openai>=0.3.11
langchain-anthropic>=0.3.10
langchain-community>=0.3.20
langchain-core>=0.3.49
langchain-text-splitters>=0.3.7
langsmith>=0.3.19

# Testing
pytest-asyncio>=0.21.0

# Development
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0